Fix bug in pythreejs OrbitControls causing rendering not to function in a jupyter notebook.
Add pyvista as a potential backend renderer and start refactor the rendering pipeline with a new api.
