+++ conda/meta.yaml
@@ -34,6 +34,7 @@ requirements:
     - python-kaleido
     - ipygany
     - pydantic
+    - pyvista
 
 about:
   home: https://github.com/krande/adapy
