+++ tests/test_viz_new_api.py
@@ -11,7 +11,7 @@ class VizApiV2(unittest.TestCase):
     def test_viz_beam(self):
         viz = Visualize()
         viz.add_obj(self.bm1)
-        viz.display()
+        viz.display(off_screen_file="temp/MyTest.svg")
 
 
 if __name__ == "__main__":
